% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.r
\name{amazon_ranges}
\alias{amazon_ranges}
\title{Amazon AWS cloud ranges}
\usage{
amazon_ranges()
}
\value{
list of 4 slots the most interesting of which is \code{prefixes} which
        is a \code{tibble} with three columns, the most interesting of which is
        \code{ip_prefix}, but the other columns are useful for various tasks. There is
        an IPv6 counterpart slot \code{ipv6_prefixes} with a similar sub-structure.
}
\description{
Amazon AWS cloud ranges
}
\examples{
ranges <- amazon_ranges()

normalize_ipv4(ranges)
}

