% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.r
\name{azure_ranges}
\alias{azure_ranges}
\title{Azure ranges}
\usage{
azure_ranges()
}
\value{
\code{data.frame} with the cloud \code{region} and cidr \code{subnet}. As of
        2016-10-04, these subnets are IPv4 only.
}
\description{
Azure ranges
}
\examples{
ranges <- azure_ranges()

normalize_ipv4(ranges)
}

