% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudcidrs-package.R
\docType{package}
\name{cloudcidrs}
\alias{cloudcidrs}
\alias{cloudcidrs-package}
\title{Tools to Obtain and Work with Cloud Provider CIDR Blocks}
\description{
Tools to Obtain and Work with Cloud Provider CIDR Blocks
}
\author{
Bob Rudis (bob@rud.is)
}

