% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linode-ranges.R
\name{linode_ranges}
\alias{linode_ranges}
\title{Linode ranges}
\usage{
linode_ranges()
}
\value{
character vector of IPv4 ranges
}
\description{
Linode (cowardly) doesn't publish their ranges, so we have to approximate with
known Linode ASNs
}
\details{
It is unlikely that this list will change in your analysis session, so it is
recommended that you cache the results. Future versions will automatically cache
the results both in-session and on-disk for a period of time.
}
\note{
ASN list current as of April, 2017
}
\examples{
ranges <- linode_ranges()

normalize_ipv4(ranges)
}
\references{
\url{http://bgp.he.net/search?search\%5Bsearch\%5D=\%linode\%22&commit=Search}
}
