% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\name{normalize_ipv4}
\alias{normalize_ipv4}
\title{Normalize Cloud CIDR return values}
\usage{
normalize_ipv4(x)
}
\arguments{
\item{x}{the output of one of the cloud \code{_ranges} functions}
}
\description{
Calling the cloud ranges functions directly returns raw data from the
various platform responses. This function will return a flat character vector
of IPv4 CIDR blocks from a cloud provider response so it can be more easily used
in downstream analysis operations
}
\examples{
amazon_ranges() \%>\% normalize_ipv4()
azure_ranges() \%>\% normalize_ipv4()
digitalocean_ranges() \%>\% normalize_ipv4()
google_ranges() \%>\% normalize_ipv4()
racksapce_ranges() \%>\% normalize_ipv4()
softlayer_ranges() \%>\% normalize_ipv4()
}

