% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovh-ranges.R
\name{ovh_ranges}
\alias{ovh_ranges}
\title{OVH ranges}
\usage{
ovh_ranges()
}
\value{
character vector of IPv4 ranges
}
\description{
OVH (cowardly) doesn't publish their ranges, so we have to approximate with
known OVH ASNs
}
\details{
It is unlikely that this list will change in your analysis session, so it is
recommended that you cache the results. Future versions will automatically cache
the results both in-session and on-disk for a period of time.
}
\note{
ASN list current as of April, 2017
}
\examples{
ranges <- ovh_ranges()

normalize_ipv4(ranges)
}
