% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.r
\name{rackspace_ranges}
\alias{rackspace_ranges}
\title{Rackspace ranges}
\usage{
rackspace_ranges()
}
\value{
character vector of IPv4 ranges
}
\description{
Rackspace (cowardly) doesn't publish their ranges, so we have to approximate with
known Rackspace Hosting ASNs
}
\note{
ASN list current as of October, 2016
}
\examples{
ranges <- rackspace_ranges()

normalize_ipvr(ranges)
}
\references{
\url{http://bgp.he.net/search?search\%5Bsearch\%5D=\%22Rackspace\%22&commit=Search}
}

