% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softlayer-ranges.R
\name{softlayer_ranges}
\alias{softlayer_ranges}
\title{Softlayer ranges}
\usage{
softlayer_ranges(method = c("asn", "list"))
}
\arguments{
\item{method}{if \code{list}, this method will use the HTML published ranges; if
\code{asn}, this method will build the CIDR list from Softlayer published
ASNs. The default method is "\code{asn}".}
}
\value{
a \code{tibble}, the most interesting colun of  which is \code{ip_range}
}
\description{
Retrieves the official list of Softlayer cloud network ranges.
}
\details{
Softlayer provides \href{https://github.com/IBM-Bluemix-Docs/hardware-firewall-dedicated/blob/master/ips.md}{a list}
of public netblock ranges. Softlayer also has a large number of large ASNs but
Softlayer tech support claimes AS36351 is their "public cloud" ASN.
Methods are provided that enable using either of these sources to generate the CIDR list.

It is unlikely that this list will change in your analysis session, so it is
recommended that you cache the results. Future versions will automatically cache
the results both in-session and on-disk for a period of time.
}
\examples{
ranges <- softlayer_ranges()

normalize_ipv4(ranges)
}
