% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.r
\name{azure_ranges}
\alias{azure_ranges}
\title{Azure ranges}
\usage{
azure_ranges()
}
\value{
\code{data.frame} with the cloud \code{region} and cidr \code{subnet}. As of
        2016-10-04, these subnets are IPv4 only.
}
\description{
Retrieves the official list of Azure cloud network ranges.
}
\details{
It is unlikely that this list will change in your analysis session, so it is
recommended that you cache the results. Future version will automatically cache
the results both in-session and on-disk for a period of time.
}
\examples{
ranges <- azure_ranges()

normalize_ipv4(ranges)
}

