% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.r
\name{digitalocean_ranges}
\alias{digitalocean_ranges}
\title{Digital Ocean ranges}
\usage{
digitalocean_ranges()
}
\value{
character vector of IPv4 ranges
}
\description{
Digitial Ocean (cowardly) doesn't publish their ranges, so we have to approximate with
known Digital Ocean ASNs.
}
\note{
ASN list current as of October, 2016
}
\examples{
ranges <- digitialocean_ranges()

normalize_ipvr(ranges)
}
\references{
\url{http://bgp.he.net/search?search\%5Bsearch\%5D=\%22Digital+Ocean\%22&commit=Search}
}

