% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.r
\name{google_ranges}
\alias{google_ranges}
\title{Google Cloud ranges}
\usage{
google_ranges()
}
\value{
list of 2 slots, one for \code{ipv4} and one for \code{ipv6} ranges
}
\description{
Google Cloud ranges
}
\examples{
ranges <- google_ranges()

normalize_ipv4(ranges)
}

