% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.r
\name{softlayer_ranges}
\alias{softlayer_ranges}
\title{Softlayer ranges}
\usage{
softlayer_ranges(method = "list")
}
\arguments{
\item{method}{if \code{list}, this method will use the HTML published ranges; if
\code{asn}, this method will build the CIDR list from Softlayer publishedf
ASNs.}
}
\value{
a \code{tibble}, the most interesting colun of  which is \code{ip_range}
}
\description{
Retrieves the official list of Softlayer cloud network ranges.
}
\details{
Softlayer provides \href{https://knowledgelayer.softlayer.com/content/what-ip-ranges-do-i-allow-through-firewall}{a list}
of public netblock ranges

It is unlikely that this list will change in your analysis session, so it is
recommended that you cache the results. Future version will automatically cache
the results both in-session and on-disk for a period of time.
}
\examples{
ranges <- softlayer_ranges()

normalize_ipv4(ranges)
}

