% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.r
\name{softlayer_ranges}
\alias{softlayer_ranges}
\title{Softlayer ranges}
\usage{
softlayer_ranges()
}
\value{
a \code{tibble}, the most interesting colun of  which is \code{ip_range}
}
\description{
Retrieves the official list of Softlayer cloud network ranges.
}
\details{
It is unlikely that this list will change in your analysis session, so it is
recommended that you cache the results. Future version will automatically cache
the results both in-session and on-disk for a period of time.
}
\examples{
ranges <- softlayer_ranges()

normalize_ipv4(ranges)
}

