% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.r
\name{softlayer_ranges}
\alias{softlayer_ranges}
\title{Softlayer ranges}
\usage{
softlayer_ranges()
}
\value{
a \code{tibble}, the most interesting colun of  which is \code{ip_range}
}
\description{
Softlayer ranges
}
\examples{
ranges <- softlayer_ranges()

normalize_ipv4(ranges)
}

